requirements.txt
mss~=9.0.1
pytesseract~=0.3.10
pillow~=10.4.0